Initial commit
Setting up the repo for fonts. This will be where I store all of the fonts we use at the library, it will make it easier to move them around and deploy them to different comptuers. This will also make it easier to update the new ones on new computers.