Added in the fonts
Added in to the repo all of the font files we are using. This is really a fork of https://github.com/Findarato/dotFiles but I am kind of probably doing it wrong.  I am pulling out all of the look And Feel into its own repo so it will be easier to clone.